pyproject.toml
[tool.poetry]
name = "уelectricmeterbot"
version = "0.1.0"
description = ""
authors = ["amantur02 <amantur.ubaid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
uvicorn = "^0.25.0"
gunicorn = "^21.2.0"
python-telegram-bot = "^20.7"
pydantic = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
